requirements.txt
APScheduler==3.6.3
CacheControl==0.12.11
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.1
firebase-admin==5.2.0
google-api-core==2.8.2
google-api-python-client==2.57.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.1
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
idna==3.3
msgpack==1.0.4
proto-plus==1.22.0
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.20.0
python-telegram-bot==13.13
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rsa==4.9
six==1.16.0
telegram==0.0.1
tornado==6.2
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.11

pandas~=1.4.3
gspread~=5.4.0